# lib/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Agendamento v2",
    "version": "1.0.0",
    "description": "API REST completa para sistema de agendamentos com autenticação via API Key",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key no formato: sk_<apiClientId>_<random>"
      },
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token do Supabase (apenas para endpoints administrativos)"
      }
    },
    "schemas": {
...
      "Professional": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/professionals": {
      "get": {
        "tags": ["Agendamentos"],
        "summary": "Listar Professionals",
        "description": "Lista todos os professionals da company associada à API Key.\n\n**Exemplo de requisição:**\n```bash\ncurl -X GET http://localhost:3000/api/v1/professionals \\\n  -H \"Authorization: Bearer <API_KEY>\" \\\n  -H \"Content-Type: application/json\"\n```",
        "security": [{"ApiKeyAuth": []}],
        "responses": {
          "200": {
            "description": "Lista de professionals retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Professional"}
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "name": "Dr. João Silva",
                      "email": "joao@example.com",
                      "phone": "+5511999999999",
                      "created_at": "2024-01-01T00:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "API Key inválida ou ausente",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Unauthorized: Invalid or missing API key"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Agendamentos"],
        "summary": "Criar Professional",
        "description": "Cria um novo professional na company associada à API Key.\n\n**Exemplo de requisição:**\n```bash\ncurl -X POST http://localhost:3000/api/v1/professionals \\\n  -H \"Authorization: Bearer <API_KEY>\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\n    \"name\": \"Dr. João Silva\",\n    \"email\": \"joao@example.com\",\n    \"phone\": \"+5511999999999\"\n  }'\n```",
        "security": [{"ApiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Dr. João Silva"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "nullable": true,
                    "example": "joao@example.com"
                  },
                  "phone": {
                    "type": "string",
                    "nullable": true,
                    "example": "+5511999999999"
                  }
                }
              },
              "example": {
                "name": "Dr. João Silva",
                "email": "joao@example.com",
                "phone": "+5511999999999"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Professional criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {"$ref": "#/components/schemas/Professional"}
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Dr. João Silva",
                    "email": "joao@example.com",
                    "phone": "+5511999999999",
                    "created_at": "2024-01-01T00:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Validation error",
                  "errors": {
                    "name": ["Required"],
                    "email": ["Invalid email"]
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key inválida ou ausente",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Unauthorized: Invalid or missing API key"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}